requirements.txt
absl-py==0.7.1
aenum==2.2.1
aiohttp==3.5.4
altgraph==0.16.1
arcade==2.1.7
asgi-redis==1.4.3
asgiref==1.1.2
astetik==1.9.8
astor==0.8.0
async-timeout==3.0.1
attrs==19.1.0
autobahn==20.3.1
Automat==20.2.0
backcall==0.1.0
beautifulsoup4==4.8.0
bleach==3.1.0
certifi==2019.3.9
cffi==1.14.0
chances==0.1.6
channels==2.4.0
chardet==3.0.4
Click==7.0
colorama==0.4.1
constantly==15.1.0
cryptography==2.8
cycler==0.10.0
daphne==2.4.1
decorator==4.4.0
defusedxml==0.6.0
discord==1.0.1
discord.py==1.1.1
dj-database-url==0.5.0
Django==3.0.1
django-heroku==0.3.1
djangorestframework==3.11.0
docutils==0.15.2
dpkt==1.9.2
entrypoints==0.3
fbchat==1.7.3
Flask==1.1.1
future==0.17.1
gast==0.2.2
geonamescache==1.0.2
grpcio==1.21.1
gunicorn==20.0.4
gym==0.12.5
h5py==2.9.0
html5lib==1.0.1
hyperlink==19.0.0
idna==2.8
idna-ssl==1.1.0
incremental==17.5.0
instagram-scraper==1.7.1
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kerasplotlib==0.1.4
Kivy==1.11.1
kivy-deps.angle==0.1.9
kivy-deps.glew==0.1.12
kivy-deps.gstreamer==0.1.17
kivy-deps.sdl2==0.1.22
Kivy-Garden==0.1.4
kiwisolver==1.1.0
leagueoflegends==1.4
lightgbm==2.2.3
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==2.2.3
mechanize==0.4.5
mistune==0.8.4
mock==3.0.5
MouseInfo==0.1.2
msgpack-python==0.5.6
multidict==4.5.2
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
numpy==1.16.3
oauthlib==3.0.1
opencv-python==4.1.0.25
pandas==0.24.2
pandocfilters==1.4.2
parso==0.4.0
patsy==0.5.1
pefile==2019.4.18
pickleshare==0.7.5
Pillow==6.0.0
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.8.0
psycopg2==2.8.4
py2exe==0.9.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.48
pycparser==2.20
PyGetWindow==0.0.7
pyglet==1.3.2
pyglet-ffmpeg2==0.1.15
Pygments==2.4.1
PyHamcrest==2.0.2
PyInstaller==3.5
PyMsgBox==1.0.7
pyOpenSSL==19.1.0
pyparsing==2.4.0
pyperclip==1.7.0
pypiwin32==223
PyQt5==5.11.3
PyQt5-sip==4.19.18
pyqt5-tools==5.11.3.1.4
PyRect==0.1.4
pyrsistent==0.15.2
PyScreeze==0.1.25
PySocks==1.7.0
python-dateutil==2.8.0
python-dotenv==0.10.3
python-nmap==0.6.1
pytiled-parser==0.9.0
PyTweening==1.0.3
pytz==2019.1
pywin32==224
pywin32-ctypes==0.2.0
pywinpty==0.5.5
PyYAML==5.1.1
pyzmq==18.0.1
qtconsole==4.5.0
redis==2.10.6
requests==2.22.0
requests-oauthlib==1.2.0
riot-observer==1.3.3
scapy==2.4.3
scapy-http==1.8.2
scikit-learn==0.21.2
scipy==1.3.0
seaborn==0.9.0
Send2Trash==1.5.0
service-identity==18.1.0
six==1.12.0
sklearn==0.0
soupsieve==1.9.2
sqlparse==0.3.0
statsmodels==0.9.0
talos==0.4.9
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tkintertable==1.3.2
tornado==6.0.2
tqdm==4.32.1
traitlets==4.3.2
tweepy==3.7.0
Twisted==19.10.0
txaio==20.1.1
typing-extensions==3.7.2
urllib3==1.25.3
virtualenv==16.6.0
virtualenvwrapper-win==1.2.5
wcwidth==0.1.7
webencodings==0.5.1
websockets==6.0
Werkzeug==0.15.4
whitenoise==5.0.1
widgetsnbextension==3.4.2
wrangle==0.6.4
xlrd==1.2.0
yarl==1.3.0
zope.interface==4.7.2
